rm log file

+++ server.py
@@ -35,6 +35,7 @@ def index():
             with open('_pre.log_') as f:
                 content = f.readlines()
             printerr = [x.strip() for x in content] 
+            rmlog() # remove log file before session ends
             return flask.render_template("outputmsg.html", msglist = printerr)
         
         res = "" # path for output file to download
@@ -48,29 +49,36 @@ def index():
         except Exception as e:
 
             return flask.render_template("outputmsg.html", msglist = messagelist[1:])
-        
+
     return render_template('index.html')
 
 def cleanUploadFolder():
         fp = os.path.join(os.path.dirname(app.instance_path), "uploads/")
         
         for file in os.listdir(fp):
+            if file != '.gitignore': # disregard the gitignore file, commit and push the empty folder uploads/
+                f_del = (os.path.join(fp, file)) 
 
-            f_del = (os.path.join(fp, file))
-            try:
-                if os.path.isfile(f_del):
-                    os.unlink(f_del)
-                elif os.path.isdir(f_del): 
-                    shutil.rmtree(f_del)
-            except Exception as e:
-                print(e)
-
+                try:
+                    if os.path.isfile(f_del):
+                        os.unlink(f_del)
+                    elif os.path.isdir(f_del): 
+                        shutil.rmtree(f_del)
+                except Exception as e:
+                    print(e)
+# create an empty log file
 def clearLogfile():
     
     filename = "_pre.log_"
     filepath = os.path.join(os.path.dirname(app.instance_path), filename)
     with open(filepath, 'w') as f:
         pass
+# remove the log file if exist
+def rmlog():
+    filepath = os.path.join(os.path.dirname(app.instance_path), "_pre.log_")
+    if os.path.exists(filepath):
+        print "here"
+        os.remove(filepath)
 
 @app.route('/about')
 def show_about():
@@ -79,7 +87,6 @@ def show_about():
 @app.route('/demo')
 def demo():
     return render_template('demo.html')
-
 
 
 if __name__=='__main__':
